adaptando projeto para maven
