+++ src/main/java/br/cefet/SliClickServerV2.java
@@ -1,0 +1,32 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package sliclickserverv2;
+
+import Serial.SerialRxTx;
+
+
+/**
+ *
+ * @author User
+ */
+public class SliClickServerV2 {
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String[] args) {
+        SerialRxTx serial = new SerialRxTx();
+        
+        if (serial.iniciaSerial()){
+            while(true){
+
+            }
+        } else {
+            
+        }
+        
+    }
+}
